Refactor: moving help into its own class, consistent final usage, version into a method
